Move functions into dedicated files
